REFACTOR: removed util.py, as it was a bit messy
