+++ mycroft/client/wifisetup/app/util/Server.py
@@ -23,7 +23,7 @@ import time
 import threading
 from Queue import Queue
 from mycroft.configuration import ConfigurationManager
-from mycroft.client.wifisetup.app.util.util import WpaClientTools, ScanForAP
+from mycroft.client.wifisetup.app.util.api import WpaClientTools, ScanForAP
 from mycroft.util.log import getLogger
 from mycroft.client.wifisetup.app.util.api import WiFiAPI, ApAPI, LinkAPI
